+++ packages/app/src/lib/components/editor.svelte
@@ -3,12 +3,7 @@
 	import {onMount} from 'svelte';
 	import cx from 'clsx';
 	import {get, type Readable} from 'svelte/store';
-	import {
-		BubbleMenu,
-		createEditor,
-		Editor,
-		EditorContent,
-	} from 'svelte-tiptap';
+	import {BubbleMenu, createEditor, Editor, EditorContent} from 'svelte-tiptap';
 	import {Collaboration} from '@tiptap/extension-collaboration';
 	import {CollaborationCursor} from '@tiptap/extension-collaboration-cursor';
 	import Placeholder from '@tiptap/extension-placeholder';
@@ -264,7 +259,7 @@
 			}
 
 			p {
-				word-break: break-all;
+				word-break: normal;
 			}
 
 			a {